Move dummy dependency fixtures to conftest
