+++ tests/test_parse_numeric_value.py
@@ -1,48 +1,7 @@
-import sys
-import types
 import pytest
-
-# Provide lightweight stubs if dependencies are missing
-if "pytz" not in sys.modules:
-    tz_module = types.ModuleType("pytz")
-
-    class DummyTZInfo:
-        def utcoffset(self, dt):
-            return None
-
-        def dst(self, dt):
-            return None
-
-        def tzname(self, dt):
-            return "UTC"
-
-        def localize(self, dt_obj):
-            return dt_obj.replace(tzinfo=self)
-
-    tz_module.timezone = lambda name: DummyTZInfo()
-    sys.modules["pytz"] = tz_module
-
-if "requests" not in sys.modules:
-    req_module = types.ModuleType("requests")
-
-    class DummySession:
-        def get(self, *args, **kwargs):
-            raise NotImplementedError
-
-    req_module.Session = DummySession
-    req_module.exceptions = types.SimpleNamespace(Timeout=Exception, ConnectionError=Exception)
-    sys.modules["requests"] = req_module
-
-if "bs4" not in sys.modules:
-    bs4_module = types.ModuleType("bs4")
-
-    class DummySoup:
-        pass
-
-    bs4_module.BeautifulSoup = DummySoup
-    sys.modules["bs4"] = bs4_module
-
 from notification_service import NotificationService
+
+pytestmark = pytest.mark.usefixtures("dummy_deps")
 
 
 class DummyState:
